+++ miracle/scripts/block.py
@@ -7,7 +7,10 @@ from sqlalchemy import (
     select,
 )
 
-from miracle.bloom import parse_domain_blocklist_source
+from miracle.bloom import (
+    create_bloom_domain,
+    parse_domain_blocklist_source,
+)
 from miracle.config import BLOOM_DOMAIN_SOURCE
 from miracle.db import create_db
 from miracle.log import (
@@ -17,34 +20,33 @@ from miracle.log import (
 from miracle.models import URL
 
 
-def remove_urls(db, lines):
+def remove_urls(db, bloom, lines):
     found_url_ids = []
     with db.session(commit=False) as session:
         # Get a list of URL ids to delete in larger batches.
-        for i in range(0, len(lines), 100):
-            name_batch = lines[i:i + 100]
-            if name_batch:
-                rows = session.execute(
-                    select([URL.id]).where(
-                        URL.hostname.in_(name_batch))).fetchall()
-                if rows:
-                    found_url_ids.extend([row[0] for row in rows])
+        for line in lines:
+            rows = session.execute(
+                select([URL.id, URL.hostname]).where(
+                    URL.hostname.endswith(line))).fetchall()
+            for row in rows:
+                if row.hostname in bloom:
+                    found_url_ids.append(row.id)
 
-    LOGGER.info('Found %s domains in database.', len(found_url_ids))
+    LOGGER.info('Found %s URLs in database.', len(found_url_ids))
     if not found_url_ids:
         return 0
 
     actually_deleted = 0
     found_url_ids.sort()
     batches = ((len(found_url_ids) - 1) // 10) + 1
-    LOGGER.info('Deleting domains in batches of 10.')
+    LOGGER.info('Deleting URLs in batches of 10.')
     for i in range(0, len(found_url_ids), 10):
         # Delete URLs by id in small transactional batches, as these
         # can result in large numbers of sessions to be deleted at the
         # same time.
         id_batch = found_url_ids[i:i + 10]
         if id_batch:
-            LOGGER.info('Deleting domains, batch %s of %s.',
+            LOGGER.info('Deleting URLs, batch %s of %s.',
                         i // 10 + 1, batches)
             with db.session() as session:
                 res = session.execute(delete(URL).where(URL.id.in_(id_batch)))
@@ -53,12 +55,12 @@ def remove_urls(db, lines):
     return actually_deleted
 
 
-def main(db, filename=BLOOM_DOMAIN_SOURCE):
+def main(db, bloom, filename=BLOOM_DOMAIN_SOURCE):
     LOGGER.info('Reading source file: %s', filename)
     lines = parse_domain_blocklist_source(filename)
     LOGGER.info('Found %s domains in source file.', len(lines))
-    urls_removed = remove_urls(db, lines)
-    LOGGER.info('Deleted %s domains.', urls_removed)
+    urls_removed = remove_urls(db, bloom, lines)
+    LOGGER.info('Deleted %s URLs.', urls_removed)
     return urls_removed
 
 
@@ -81,6 +83,8 @@ def console_entry():  # pragma: no cover
 
     try:
         db = create_db()
-        main(db, filename)
+        bloom = create_bloom_domain()
+        main(db, bloom, filename)
     finally:
+        bloom.close()
         db.close()