+++ miracle/scripts/tests/test_block.py
@@ -11,14 +11,17 @@ from miracle.scripts import block
 TEST_START = datetime.utcfromtimestamp(1469400000)
 
 
-def test_remove_urls(db):
+def test_remove_urls(db, bloom_domain):
     # Test without any matching URLs
-    assert block.remove_urls(db, ['localhost']) == 0
+    assert block.remove_urls(db, bloom_domain, ['localhost']) == 0
 
     with db.session() as session:
         url1 = URL(**URL.from_url('http://localhost:80/path'))
         url2 = URL(**URL.from_url('https://localhost/'))
         url3 = URL(**URL.from_url('https://example.com/'))
+        url4 = URL(**URL.from_url('https://localhost.local/'))
+        url5 = URL(**URL.from_url('https://b.localhost.local/'))
+        url6 = URL(**URL.from_url('https://a.b.localhost.local/'))
         user1 = User(token='user1')
         user2 = User(token='user2')
         user3 = User(token='user3')
@@ -26,10 +29,14 @@ def test_remove_urls(db):
             Session(user=user1, url=url1, start_time=TEST_START),
             Session(user=user1, url=url2, start_time=TEST_START),
             Session(user=user1, url=url3, start_time=TEST_START),
+            Session(user=user1, url=url4, start_time=TEST_START),
+            Session(user=user1, url=url5, start_time=TEST_START),
+            Session(user=user1, url=url6, start_time=TEST_START),
             Session(user=user2, url=url1, start_time=TEST_START),
             Session(user=user3, url=url3, start_time=TEST_START),
         ])
-    assert block.remove_urls(db, ['localhost']) == 2
+    assert block.remove_urls(
+        db, bloom_domain, ['localhost', 'localhost.local']) == 5
 
     with db.session() as session:
         hostnames = {row[0] for row in session.query(URL.hostname).all()}
@@ -40,9 +47,9 @@ def test_remove_urls(db):
             {('user1', 'example.com'), ('user3', 'example.com')}
 
 
-def test_main(db, tmp_path):
+def test_main(db, bloom_domain, tmp_path):
     in_filename = os.path.join(tmp_path, 'bloom.txt')
     with open(in_filename, 'wt') as fd:
         fd.write('localhost\n')
 
-    assert block.main(db, filename=in_filename) == 0
+    assert block.main(db, bloom_domain, filename=in_filename) == 0