+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 1.1.4 (unreleased)
 ==================
 
+- Add public_suffix list (2016-09-14).
+
 - Extend domain blocklist.